caddy TLS test progress v1.2

+++ ci/create-link-ci.sh
@@ -10,44 +10,34 @@ docker compose up -d --build
 eval $(ssh-agent -s)
 ssh-add ./gateway-sim-key
 
-# Test create-link
-# generate a docker compose to test the generated link
-cat test-link.template.yaml > test-link.yaml
-docker run --network gateway -e SSH_AGENT_PID=$SSH_AGENT_PID -e SSH_AUTH_SOCK=$SSH_AUTH_SOCK -v $SSH_AUTH_SOCK:$SSH_AUTH_SOCK --rm fractalnetworks/gateway-cli:latest $1 $2 $3 >> test-link.yaml
-cat network.yaml >> test-link.yaml
-# set the gateway endpoint to the gateway link container
-sed -i 's/^\(\s*GATEWAY_ENDPOINT:\).*/\1 app-example-com:18521/' test-link.yaml
 
-docker compose -f test-link.yaml up -d
-docker compose -f test-link.yaml exec link ping 10.0.0.1 -c 2
-# assert http response code was 200
-# asserts basic auth is working with user: admin, password: admin
+testLinkFile=""   # Define the variable in a scope outside the cleanup function
 
-if ! docker compose exec gateway curl -k -H "Authorization: Basic YWRtaW46YWRtaW4=" --resolve app.example.com:443:127.0.0.1 https://app.example.com -I |grep "HTTP/2 200"; then
-    FAILED="true"
-fi
-docker compose -f test-link.yaml down
-docker rm -f app-example-com
-rm test-link.yaml                 # comment out to keep the file for debugging
+function cleanup {
+    if [[ -n "$testLinkFile" ]]; then  # Check if the variable is non-empty
+        echo "\n******* Cleanup function: cleaning up $testLinkFile..."
+        docker compose -f "$testLinkFile" down --remove-orphans || true
+        docker rm -f app-example-com || true
+
+        rm "$testLinkFile" || true        # comment out to keep the file for debugging
+    fi
+}
+# Catch and cleanup stragglers if the script fails or is terminated.
+# Good for local testing, eliminates the need to manually remove docker containers.
+trap cleanup EXIT
 
 
-caddy_greenlight=true               # andrew's sentinel thing
+normal_test_greenlight=false               # andrew's sentinel thing
+if [ "$normal_test_greenlight" = true ]; then
+    # Test create-link
+    $testLinkFile="test-link.yaml"
 
-if [ "$caddy_greenlight" = true ]; then
-    echo "*******************Testing Caddy TLS Proxy"
-    # Test the link using  CADDY_TLS_PROXY: true
-    # generate new docker compose
-    testLinkFile="test-link-caddyTLS.yaml"
+    # generate a docker compose to test the generated link
     cat test-link.template.yaml > $testLinkFile
     docker run --network gateway -e SSH_AGENT_PID=$SSH_AGENT_PID -e SSH_AUTH_SOCK=$SSH_AUTH_SOCK -v $SSH_AUTH_SOCK:$SSH_AUTH_SOCK --rm fractalnetworks/gateway-cli:latest $1 $2 $3 >> $testLinkFile
     cat network.yaml >> $testLinkFile
-
-    # Go inside $testLinkFile and change...
-    # gateway endpoint to the gateway link container
+    # set the gateway endpoint to the gateway link container
     sed -i 's/^\(\s*GATEWAY_ENDPOINT:\).*/\1 app-example-com:18521/' $testLinkFile
-
-    # set CADDY_TLS_PROXY to true
-    sed -i 's/^\(\s*\)#\s*CADDY_TLS_PROXY: true/\1CADDY_TLS_PROXY: true/' $testLinkFile
 
     docker compose -f $testLinkFile up -d
     docker compose -f $testLinkFile exec link ping 10.0.0.1 -c 2
@@ -57,14 +47,61 @@ if [ "$caddy_greenlight" = true ]; then
     if ! docker compose exec gateway curl -k -H "Authorization: Basic YWRtaW46YWRtaW4=" --resolve app.example.com:443:127.0.0.1 https://app.example.com -I |grep "HTTP/2 200"; then
         FAILED="true"
     fi
+
+    # cleanup
     docker compose -f $testLinkFile down
+    docker rm -f app-example-com
+    rm $testLinkFile               # comment out to keep the file for debugging
+else
+    echo "******************* Skipping normal link test... (normal_test_greenlight was false)"
+fi
+
+
+caddy_greenlight=true               # andrew's sentinel thing
+
+if [ "$caddy_greenlight" = true ]; then
+    echo "******************* Testing Caddy TLS Proxy *******************"
+    # Test the link using  CADDY_TLS_PROXY: true
+    testLinkFile="test-link-caddyTLS.yaml"
+
+    # generate new docker compose
+    cat test-link.template.yaml > $testLinkFile
+    docker run --network gateway -e SSH_AGENT_PID=$SSH_AGENT_PID -e SSH_AUTH_SOCK=$SSH_AUTH_SOCK -v $SSH_AUTH_SOCK:$SSH_AUTH_SOCK --rm fractalnetworks/gateway-cli:latest $1 $2 $3 >> $testLinkFile
+    cat network.yaml >> $testLinkFile
+
+    # Go inside $testLinkFile and change... (requires the commented options to be there! Can change later)
+    # 1. gateway endpoint to the gateway link container
+    sed -i 's/^\(\s*GATEWAY_ENDPOINT:\).*/\1 app-example-com:18521/' $testLinkFile
+
+    # 2. CADDY_TLS_PROXY to ------------------------------------- true
+    sed -i 's/^\(\s*\)#\s*CADDY_TLS_PROXY: true/\1CADDY_TLS_PROXY: true/' $testLinkFile
+
+    # 3. CADDY_TLS_INSECURE to ---------------------------------------- false
+    sed -i 's/^\(\s*\)#\s*CADDY_TLS_INSECURE: true/\1CADDY_TLS_INSECURE: false/' $testLinkFile
+
+    # 4. FORWARD_ONLY to ----------------------------------- false
+    # sed -i 's/^\(\s*\)#\s*FORWARD_ONLY: true/\1FORWARD_ONLY: false/' $testLinkFile
+        #****** works when false ******
+        # otherwise can't ping:  "Error response from daemon: container [container identifier] is not running"
+
+    # docker compose -f $testLinkFile up > "$testLinkFile"-compose-up.log 2>&1
+    docker compose -f $testLinkFile up -d
+    docker compose -f $testLinkFile exec link ping 10.0.0.1 -c 2
+    # assert http response code was 200
+    # asserts basic auth is working with user: admin, password: admin
+
+    # (??) don't use "-k" for caddy if testing CADDY_TLS_INSECURE: false (aka https)
+    if ! docker compose exec gateway curl -k -H "Authorization: Basic YWRtaW46YWRtaW4=" --resolve app.example.com:443:127.0.0.1 https://app.example.com -I 2>&1 | tee curl-output.log |grep "HTTP/2 200"; then
+        FAILED="true"
+    fi
+    docker compose -f $testLinkFile down --remove-orphans
     docker rm -f app-example-com
     rm $testLinkFile                 # comment out to keep the file for debugging
 fi
 
 
 # stop and remove gateway and sshd containers
-docker compose down
+docker compose down || echo "'docker compose down' at the end had an issue"
 
 # if FAILED is true return 1 else 0
 if [ ! -z ${FAILED+x} ]; then
